LPD-35051 - Feat. Extract FieldSelectModalComponent to its own component
